Changed Bokeh requirements to work with panel 1.x
Moved code from nbinit that checked CLI credentials to be run only in AML - avoiding non-core import.
