+++ docs/source/conf.py
@@ -62,8 +62,8 @@ extensions = [
     "sphinxcontrib.jquery",
     "sphinx.ext.napoleon",
     "sphinx.ext.autosectionlabel",
-    "sphinx.ext.intersphinx",
-    "seed_intersphinx_mapping",
+    # "sphinx.ext.intersphinx",
+    # "seed_intersphinx_mapping",
 ]
 
 autosectionlabel_prefix_document = True
