+++ msticpy/_version.py
@@ -1,2 +1,2 @@
 """Version file."""
-VERSION = "2.5.1"
+VERSION = "2.5.2"
